Fix AMI Re-connection (#62)
* Change reconnection times

The old reconnection metodology was increasing the reconnection time almost for a day, that was to much, so, now the default reconnection was change to 10 seconds. For each reconnection try, the time is incresed 10 seconds, the reconnection may reach a maximum of 60 seconds.

* Missed parameter when reconnecting

The connect function was expecting 4 parameters, but only 3 parameters were sending, due this, the host parameter was containing invalid value, avoiding to recconect to Asterisk AMI.

* Remove Node Reference

Due the author of this plugin doesn't update the node repo, I think is not necessary to show the reference to node PKG, actually the node code is old and have not been updated for awhile.